refactor(daemon): deduplicate Wayland and X11 main loops (P-08)
- Extract shared main loop state into `MainLoopContext` in `main_loop.rs`
- Move shared logic (draining channels, scheduling, commands) to `MainLoopContext` methods
- Separate backend-specific logic into `wayland_loop.rs` and `x11_loop.rs`
- Slim `main.rs` to a thin dispatcher for args, config, and backend selection
- Reduce total loop code by ~700 lines through shared abstractions
